+++ homework8/src/main/java/by/popkov/homework8/weather_api_data_classes/WeatherApiForecast.java
@@ -1,14 +1,14 @@
-package by.popkov.homework8;
+package by.popkov.homework8.weather_api_data_classes;
 
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
 
-class WeatherApiForecast {
+public class WeatherApiForecast {
     @SerializedName("list")
     private List<WeatherApiForecastListObject> weatherApiForecastList;
 
-    List<WeatherApiForecastListObject> getWeatherApiForecastList() {
+    public List<WeatherApiForecastListObject> getWeatherApiForecastList() {
         return weatherApiForecastList;
     }
 }
